Added the KeyListenerBehaviour
The KeyListenerBehaviour listens to user input from the keyboard.
Basicly what the PlayerFish used to do. Now the playerFish simply has
the KeyListenerBehaviour as an attribute like every Entity.
