+++ src/main/java/com/github/fishio/PlayerFish.java
@@ -3,30 +3,19 @@ package com.github.fishio;
 import com.github.fishio.achievements.State;
 import com.github.fishio.achievements.Subject;
 import com.github.fishio.behaviours.IBehaviour;
+import com.github.fishio.behaviours.KeyListenerBehaviour;
 
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 import javafx.scene.input.KeyCode;
-import javafx.scene.input.KeyEvent;
 import javafx.stage.Stage;
 
 /**
  * Represents a fish that the user can control using
  * the keyboard.
  */
-public class PlayerFish extends Entity implements IEatable, IPositional, IBehaviour, Subject {
-
-	private double vx;
-	private double vy;
-
-	/**
-	 * These factors have values for whether each of the arrow keys is pressed.
-	 */
-	private boolean upPressed;
-	private boolean downPressed;
-	private boolean leftPressed;
-	private boolean rightPressed;
+public class PlayerFish extends Entity implements IEatable, IPositional, Subject {
 
 	private static final KeyCode KEY_UP = KeyCode.UP;
 	private static final KeyCode KEY_DOWN = KeyCode.DOWN;
@@ -37,22 +26,18 @@ public class PlayerFish extends Entity implements IEatable, IPositional, IBehavi
 
 	private SimpleIntegerProperty score = new SimpleIntegerProperty(0);
 
-	/**
-	 * The speed at which the speed of the fish increases /
-	 * decreases depending on what keys are pressed by the user.
-	 */
-	private double acceleration = 0.1;
-
-	private double maxSpeed = 4;
-
 	private static final double GROWTH_SPEED = 500;
 	private static final double FISH_EAT_THRESHOLD = 1.2;
+	private static final double DEFAULT_ACCELERATION = 0.1;
+	private static final double DEFAULT_MAX_SPEED = 4;
 	private static final int START_LIVES = 3;
 	public static final int MAX_LIVES = 5;
 	
 	private SimpleIntegerProperty lives = new SimpleIntegerProperty(START_LIVES);
 	
 	private long invincible;
+	
+	private IBehaviour behaviour;
 
 	/**
 	 * Creates the Player fish which the user will be able to control.
@@ -68,202 +53,9 @@ public class PlayerFish extends Entity implements IEatable, IPositional, IBehavi
 		super(ca);		
 
 		this.sprite = sprite;
-
-		stage.addEventHandler(KeyEvent.KEY_PRESSED, event -> {
-			KeyCode pressedKey = event.getCode();
-			if (pressedKey == KEY_UP) {
-				upPressed = true;
-			} else if (pressedKey == KEY_DOWN) {
-				downPressed = true;
-			} else if (pressedKey == KEY_LEFT) {
-				leftPressed = true;
-			} else if (pressedKey == KEY_RIGHT) {
-				rightPressed = true;
-			}
-		});
-
-		stage.addEventHandler(KeyEvent.KEY_RELEASED, event -> {
-			KeyCode releasedKey = event.getCode();
-			if (releasedKey == KEY_UP) {
-				upPressed = false;
-			} else if (releasedKey == KEY_DOWN) {
-				downPressed = false;
-			} else if (releasedKey == KEY_LEFT) {
-				leftPressed = false;
-			} else if (releasedKey == KEY_RIGHT) {
-				rightPressed = false;
-			}
-		});
-	}
-
-	/**
-	 * Increases or decreases the speed of the fish in the
-	 * horizontal direction, depending on which keys are currently
-	 * pressed by the user.
-	 */
-	public void adjustXSpeed() {
-		if (leftPressed && -vx < maxSpeed) {
-			vx -= acceleration;
-		}
-		if (rightPressed && vx < maxSpeed) {
-			vx += acceleration;
-		}
-
-		if (vx < 0 && (!leftPressed || rightPressed)) {
-			vx += acceleration;
-		} else if (vx > 0 && (!rightPressed || leftPressed)) {
-			vx -= acceleration;
-		}
-
-		if (vx < 0.1 && vx > -0.1) {
-			vx = 0;	// stop if speed is too slow
-		}
-	}
-
-	/**
-	 * Increases or decreases the speed of the fish in the
-	 * vertical direction, depending on which keys are currently
-	 * pressed by the user.
-	 */
-	public void adjustYSpeed() {
-		if (downPressed && -vy < maxSpeed) {
-			vy -= acceleration;
-		}
-
-		if (upPressed && vy < maxSpeed) {
-			vy += acceleration;
-		}
-
-		if (vy < 0 && (!downPressed || upPressed)) {
-			vy += acceleration;
-		} else if (vy > 0 && (!upPressed || downPressed)) {
-			vy -= acceleration;
-		}
-
-		if (vy < 0.1 && vy > -0.1) {
-			vy = 0;	// stop if speed is too slow
-		}
-	}
-
-	/**
-	 * Gives the horizontal speed of the Player Fish.
-	 * 
-	 * @return the horizontal speed of the PlayerFish. A negative speed means
-	 *         the fish is going left.
-	 */
-	public double getSpeedX() {
-		return vx;
-	}
-
-	/**
-	 * Gives the vertical speed of the Player Fish.
-	 * 
-	 * @return the vertical speed of the PlayerFish. A negative speed means the
-	 *         fish is going down.
-	 */
-	public double getSpeedY() {
-		return vy;
-	}
-
-	/**
-	 * Gives the acceleration the Player Fish has.
-	 * 
-	 * @return the acceleration of the fish.
-	 */
-	public double getAcceleration() {
-		return acceleration;
-	}
-	
-	/**
-	 * Sets the acceleration of this PlayerFish.
-	 * 
-	 * @param acceleration
-	 * 		The new acceleration of this PlayerFish
-	 */
-	public void setAcceleration(double acceleration) {
-		this.acceleration = acceleration;
-	}
-	
-	/**
-	 * @return
-	 * 		the maximum speed of this PlayerFish
-	 */
-	public double getMaxSpeed() {
-		return maxSpeed;
-	}
-	
-	/**
-	 * Sets the maximum speed of this PlayerFish.
-	 * 
-	 * @param maxSpeed
-	 * 		The new maximum speed of this PlayerFish.
-	 */
-	public void setMaxSpeed(double maxSpeed) {
-		this.maxSpeed = maxSpeed;
-	}
-
-	/**
-	 * Sets the speed of the Player Fish in the horizontal direction.
-	 * 
-	 * @param speedX
-	 *            the speed to set
-	 */
-	public void setSpeedX(double speedX) {
-		this.vx = speedX;
-	}
-
-	/**
-	 * Sets the speed of the PlayerFish in the vertical direction.
-	 * 
-	 * @param speedY
-	 * 		the speed to set
-	 */
-	public void setSpeedY(double speedY) {
-		this.vy = speedY;
-	}
-
-	/**
-	 * Sets whether the class believes if the 
-	 * up key is pressed or not. 
-	 * 
-	 * @param upPressed
-	 * 		Whether the up key is pressed or not.
-	 */
-	public void setUpPressed(boolean upPressed) {
-		this.upPressed = upPressed;
-	}
-
-	/**
-	 * Sets whether the class believes if the 
-	 * down key is pressed or not. 
-	 * 
-	 * @param downPressed
-	 * 		Whether the down key is pressed or not.
-	 */
-	public void setDownPressed(boolean downPressed) {
-		this.downPressed = downPressed;
-	}
-
-	/**
-	 * Sets whether the class believes if the 
-	 * left key is pressed or not. 
-	 * 
-	 * @param leftPressed
-	 * 		Whether the left key is pressed or not.
-	 */
-	public void setLeftPressed(boolean leftPressed) {
-		this.leftPressed = leftPressed;
-	}
-
-	/**
-	 * Sets whether the class believes if the 
-	 * right key is pressed or not. 
-	 * 
-	 * @param rightPressed
-	 * 		Whether the right key is pressed or not.
-	 */
-	public void setRightPressed(boolean rightPressed) {
-		this.rightPressed = rightPressed;
+		
+		this.behaviour = new KeyListenerBehaviour(stage, KEY_UP, KEY_DOWN, KEY_LEFT, KEY_RIGHT,
+				DEFAULT_ACCELERATION, DEFAULT_MAX_SPEED);
 	}
 
 	/**
@@ -273,17 +65,6 @@ public class PlayerFish extends Entity implements IEatable, IPositional, IBehavi
 	 */
 	public double getGrowthSpeed() {
 		return GROWTH_SPEED;
-	}
-
-	@Override
-	public Vec2d getSpeedVector() {
-		return new Vec2d(getSpeedX(), getSpeedY());
-	}
-
-	@Override
-	public void preMove() {
-		adjustXSpeed();
-		adjustYSpeed();
 	}
 
 	@Override
@@ -368,14 +149,14 @@ public class PlayerFish extends Entity implements IEatable, IPositional, IBehavi
 		if (isDead()) {
 			return;
 		}
-		getBoundingArea().setRotation(this); //update rotation;
+		getBoundingArea().setRotation(behaviour); //update rotation;
 
-		if (vx > 0) {
+		if (behaviour.getSpeedVector().x > 0) {
 			drawRotatedImage(gc, sprite, getBoundingArea(), false);
-		} else if (vx < 0) {
+		} else if (behaviour.getSpeedVector().x < 0) {
 			drawRotatedImage(gc, sprite, getBoundingArea(), true);
 		} else {
-			drawRotatedImage(gc, sprite, getBoundingArea(), vy < 0);
+			drawRotatedImage(gc, sprite, getBoundingArea(), behaviour.getSpeedVector().y < 0);
 		}
 	}
 	
@@ -481,12 +262,12 @@ public class PlayerFish extends Entity implements IEatable, IPositional, IBehavi
 
 	@Override
 	public IBehaviour getBehaviour() {
-		return this;
+		return behaviour;
 	}
 
 	@Override
 	public void setBehaviour(IBehaviour behaviour) {
-		//TODO
+		this.behaviour = behaviour;
 	}
 
 }
